Refactor (move 'printSelectedValues' & 'printBasicInfo' to an own file)
