+++ bin/dart_kassakuitti_cli.dart
@@ -1,14 +1,12 @@
 import 'dart:io';
 
-import 'package:path/path.dart';
-import 'package:yaml/yaml.dart';
-
 import 'read_ean_products.dart';
 import 'ean_products_2_csv.dart';
 import 'specific/s_kaupat/read_receipt_products.dart';
 import 'specific/s_kaupat/receipt_products_2_csv.dart';
 import 'utils/arg_selector_helper.dart';
 import 'utils/parse_kassakuitti_arguments.dart';
+import 'utils/printing_helper.dart';
 import 'utils/shop_selector_helper.dart';
 
 void main(List<String> arguments) async {
@@ -29,19 +27,17 @@ void main(List<String> arguments) async {
         argResults[ArgSelector.foodOnlineStore.value!] as String;
     var csvFilesPath = argResults[ArgSelector.csvPath.value!] as String;
 
-    _printSelectedValues(
+    printSelectedValues(
         selectedTextFile, selectedHtmlFile, selectedStore, csvFilesPath);
 
     try {
       if (ShopSelector.sKaupat.isEqual(selectedStore)) {
         var receiptProducts =
             await readReceiptProducts(selectedTextFile, csvFilesPath);
-
-        receiptProducts2CSV(receiptProducts, csvFilesPath);
-
         var eanProducts = await readEANProducts(
             selectedHtmlFile, ShopSelector.sKaupat, csvFilesPath);
 
+        receiptProducts2CSV(receiptProducts, csvFilesPath);
         eanProducts2CSV(eanProducts, csvFilesPath, ShopSelector.sKaupat.name);
       } else if (ShopSelector.kRuoka.isEqual(selectedStore)) {
         var eanProducts = await readEANProducts(
@@ -59,36 +55,6 @@ void main(List<String> arguments) async {
 
     print('Done!');
   } else {
-    await _printBasicInfo();
+    await printBasicInfo();
   }
-}
-
-Future<void> _printBasicInfo() async {
-  String pathToYaml =
-      join(dirname(Platform.script.toFilePath()), '../pubspec.yaml');
-  var file = File(pathToYaml);
-  var fileAsString = await file.readAsString();
-
-  Map yaml = loadYaml(fileAsString);
-  print(yaml['name']);
-  print(yaml['description']);
-  print('Version: ${yaml['version']}');
-  print('Homepage: ${yaml['homepage']}');
-
-  print('''\nTo get help, run:
-  
-    dart run bin/dart_kassakuitti_cli.dart help
-    
-  or when using alias:
-    
-    kassakuitti help\n''');
-}
-
-void _printSelectedValues(String selectedTextFile, String selectedHtmlFile,
-    String selectedStore, String csvFilesPath) {
-  print('''Selected values:
-    - Path to the cash receipt:\t\t$selectedTextFile
-    - Path to the EAN products file:\t$selectedHtmlFile
-    - Selected store:\t\t\t$selectedStore
-    - Path where to save CSV files:\t$csvFilesPath''');
 }
